backend/requirements.txt
Django==4.0.0
django-filter>=23.2
djangorestframework>=3.14.0
drf-base64>=2.0
isort>=5.12.0
Pillow>=10.0.0
python-dotenv>=1.0.0
djoser>=2.2.0
pytest>=6.2.4
pytest-django>=4.4.0
pytest-pythonpath>=0.7.3
psycopg2>=2.9.9